+++ modmain.lua
@@ -105,7 +105,7 @@ modimport("scripts/enchants/youjishucai")
 modimport("scripts/enchants/qiangwei")
 modimport("scripts/enchants/dengdengqiuling")
 modimport("scripts/enchants/epsilon")
--- modimport("scripts/enchants/xinshidi")
+modimport("scripts/enchants/xinshidi")
 
 -- ------------------------------------------------------------------
 -- 4. UI 界面 (仅当任一相关功能启用时加载)